+++ tasks/index.mjs
@@ -1,0 +1,10 @@
+/**
+ * Shared tasks
+ */
+export * as assets from './assets.mjs'
+export * as browser from './browser.mjs'
+export * as configs from './configs.mjs'
+export * as files from './files.mjs'
+export * as npm from './npm.mjs'
+export * as scripts from './scripts.mjs'
+export * as styles from './styles.mjs'
